requirements.txt
pandas
numpy
tensorflow
joblib
scikit-learn
imblearn
mysql-connector-python
streamlit
pytz
